Add runnable scripts for all benchmarks
This change adds a `run-benchmark.sh` script for each of the benchmarks
used in the cluster-script setup.  These scripts each run the
corresponding benchmark and generate the necessary CSV. This removes the
need for the code to be in the Kubernetes YAML file, only the parameters
need to be passed.

For the benchmarks that have a server and a client, there's a
`run-server.sh` script that takes care of starting the server.